refactor: ladderGameData 에서 get data하지 않고 ladder 내부에서 필요한 데이터 꺼내도록 변경
