+++ src/main/java/ladder/LadderGameApplication.java
@@ -2,19 +2,18 @@ package ladder;
 
 import ladder.domain.*;
 import ladder.util.InputHelper;
-import ladder.util.LadderMaker;
 import ladder.view.InputView;
 import ladder.view.OutputView;
 
 public class LadderGameApplication {
     public static void main(String[] args) {
         LadderGameData ladderGameData = generateData();
-        Ladder ladder = new Ladder(LadderMaker.generateLadder(ladderGameData.getHeight(), ladderGameData.getPerson().getCountOfPerson()));
+        Ladder ladder = Ladder.newLadder(ladderGameData);
         OutputView.printLadder(ladder, ladderGameData);
         RewardPersonConnector rewardPersonConnector = new RewardPersonConnector(ladder, ladderGameData);
         String requestedName;
         do {
-            requestedName = InputView.findResultName(ladderGameData.getPerson());
+            requestedName = InputView.findResultName(ladderGameData);
             OutputView.printLadderResult(rewardPersonConnector.getResult(requestedName));
         } while (!InputHelper.isAll(requestedName));
     }
