+++ src/main/java/ladder/domain/Ladder.java
@@ -5,8 +5,12 @@ import java.util.List;
 public class Ladder {
     private final List<Line> lines;
 
-    public Ladder(final List<Line> lines) {
+    private Ladder(final List<Line> lines) {
         this.lines = lines;
+    }
+
+    public static Ladder newLadder(LadderGameData ladderGameData) {
+        return new Ladder(LadderMaker.generateLadder(ladderGameData));
     }
 
     int moveStartToEnd(int index) {
